requirements.txt
azure-storage-blob==12.7.1 